Nessa commit eu comecei a refatoração do código para dentro dos pacotes para aí sim criar o viewmodel da lista e o repositório. Vale lembrar que o serviço deve ser dividido mesmo tendo o mesmo url base, para os dois serviços: Lista e detalhes.
